+++ pytket/binders/passes.cpp
@@ -14,6 +14,8 @@
 
 #include <pybind11/functional.h>
 
+#include <tklog/TketLog.hpp>
+
 #include "binder_json.hpp"
 #include "tket/ArchAwareSynth/SteinerForest.hpp"
 #include "tket/Mapping/LexiLabelling.hpp"
@@ -503,9 +505,16 @@ PYBIND11_MODULE(passes, m) {
       "When merging rotations with the same op group name, the merged "
       "operation keeps the same name.");
   m.def(
-      "SynthesiseHQS", &SynthesiseHQS,
+      "SynthesiseHQS",
+      []() {
+        tket_log()->warn(
+            "SynthesiseHQS is deprecated. It will be removed "
+            "after pytket v1.25.");
+        return SynthesiseHQS();
+      },
       "Optimises and converts a circuit consisting of CX and single-qubit "
-      "gates into one containing only ZZMax, PhasedX, Rz and Phase.");
+      "gates into one containing only ZZMax, PhasedX, Rz and Phase. "
+      "DEPRECATED: will be removed after pytket 1.25.");
   m.def(
       "SynthesiseTK", &SynthesiseTK,
       "Optimises and converts all gates to TK2, TK1 and Phase gates.");